Move nameToDisplay() to mixin SubmodelElementHandling
