+++ aas-gui/Frontend/aas-web-gui/src/components/SubmodelPlugins/ContactInformation.vue
@@ -341,14 +341,6 @@ export default defineComponent({
             }
             return translatedRole;
         },
-
-        nameToDisplay(sme: any) {
-            if (sme.displayName) {
-                let displayNameEn = sme.displayName.find((displayName: any) => { return (displayName.language === 'en' && displayName.text !== ''); });
-                if (displayNameEn && displayNameEn.text) return displayNameEn.text;
-            }
-            return (sme.idShort ? sme.idShort : '');
-        },
     },
 });
 </script>
