+++ aas-gui/Frontend/aas-web-gui/src/components/SubmodelPlugins/TechnicalData.vue
@@ -278,14 +278,6 @@ export default defineComponent({
                 });
             }
         },
-
-        nameToDisplay(sme: any) {
-            if (sme.displayName) {
-                let displayNameEn = sme.displayName.find((displayName: any) => { return (displayName.language === 'en' && displayName.text !== ''); });
-                if (displayNameEn && displayNameEn.text) return displayNameEn.text;
-            }
-            return (sme.idShort ? sme.idShort : '');
-        },
     },
 });
 </script>
