+++ aas-gui/Frontend/aas-web-gui/src/components/UIComponents/IdentificationElement.vue
@@ -9,7 +9,7 @@
                 </v-tooltip>
                 <!-- idShort -->
                 <template v-slot:title>
-                    <div class="text-primary text-subtitle-1">{{ nameToDisplay }}</div>
+                    <div class="text-primary text-subtitle-1">{{ nameToDisplay(identificationObject) }}</div>
                     <div v-if="identificationObject.id">{{ idType + ':' }}</div>
                 </template>
                 <!-- identification id -->
@@ -31,9 +31,11 @@
 <script lang="ts">
 import { defineComponent } from 'vue';
 import { useNavigationStore } from '@/store/NavigationStore';
+import SubmodelElementHandling from '@/mixins/SubmodelElementHandling';
 
 export default defineComponent({
     name: 'IdentificationElement',
+    mixins: [SubmodelElementHandling],
     props: ['identificationObject', 'modelType', 'idType', 'nameType'],
 
     setup() {
@@ -48,17 +50,6 @@ export default defineComponent({
         return {
             copyIcon: 'mdi-clipboard-file-outline',
         }
-    },
-
-    computed: {
-        // Name of the identificationObject to be displayed
-        nameToDisplay() {
-            if (this.identificationObject.displayName) {
-                let displayNameEn = this.identificationObject.displayName.find((displayName: any) => { return (displayName.language === 'en' && displayName.text !== ''); });
-                if (displayNameEn && displayNameEn.text) return displayNameEn.text;
-            }
-            return (this.identificationObject.idShort ? this.identificationObject.idShort : '');
-        },
     },
 
     methods: {
