+++ aas-gui/Frontend/aas-web-gui/src/mixins/SubmodelElementHandling.ts
@@ -498,5 +498,14 @@ export default defineComponent({
             }
             return '';
         },
+
+        // Name to be displayed
+        nameToDisplay(sme: any) {
+            if (sme.displayName) {
+                let displayNameEn = sme.displayName.find((displayName: any) => { return (displayName.language === 'en' && displayName.text !== ''); });
+                if (displayNameEn && displayNameEn.text) return displayNameEn.text;
+            }
+            return (sme.idShort ? sme.idShort : '');
+        },
     },
 })
